Model sliced-cup
I was experimenting with double-side milling. For some reason, the
registration holes from Kiri:Moto didn't work for me 🤷‍♂️ so I had to
create the holes in the model.

Lessons learned: pay close attention to the material thickness (I was
short by 0.95 millimeter). Also put the register holes as farther as
possible in the material, that increases precision.