Mind blowing insight
I've started working with "Round-anything" and the workflow blew my
mind.

Openscad is based on adding and subtracting solids.
Round-anything is based on 2D scatching and extruding.

The latter approach is not only easier in general but a perfect fit for
CNC projects, which essentially works with boards and the Z axis is
always straigth.

I've tried to redo my original stool using it, and it's so much easier.
The problem is that Openscad doesn't have primitives for working with 2D
scratches. So I've started building some.

It quickly became a new set of tools and concepts, all based on
vectors (points) and lines. Things like "midpoint", or "the point
crossing 2 lines", or even "snap to a point". When I've applied that to
some of my projects, everything become easier to write, easier to
follow, and much more robust to parametric changes.

Yeah... for me it's being mind blowing.